스탬프 추가 전체적인 BigDecimal타입 int로변경

+++ src/main/java/com/example/dbcafe/member/dto/MenuDTO.java
@@ -15,7 +15,7 @@ public class MenuDTO {
 
     private String menuName;
 
-    private BigDecimal menuPrice; //String >> BigDecimal로 수정했음
+    private int menuPrice; //String >> BigDecimal로 수정했음
 
     private String menuImagePath;
